+++ flutter_full_screen_scanner_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.7
+* Bumped version to match main package.
+
 ## 1.0.6
 * Bumped version to match main package.